refactor: use a screen
