+++ src/components/BottomSheet.tsx
@@ -15,13 +15,12 @@ interface Props {
   title?: string | null
   description?: string | null
   children?: React.ReactNode | React.ReactNode[]
-  onDismiss?: () => void
   testId: string
 }
 
 export type BottomSheetRefType = GorhomBottomSheet
 
-const BottomSheet = ({ forwardedRef, title, description, children, onDismiss, testId }: Props) => {
+const BottomSheet = ({ forwardedRef, title, description, children, testId }: Props) => {
   const initialSnapPoints = useMemo(() => ['CONTENT_HEIGHT'], [])
   const { animatedHandleHeight, animatedSnapPoints, animatedContentHeight, handleContentLayout } =
     useBottomSheetDynamicSnapPoints(initialSnapPoints)
@@ -33,12 +32,6 @@ const BottomSheet = ({ forwardedRef, title, description, children, onDismiss, te
     []
   )
 
-  const handleChangePosition = (index: number) => {
-    if (index === -1) {
-      onDismiss?.()
-    }
-  }
-
   return (
     <GorhomBottomSheet
       ref={forwardedRef}
@@ -49,7 +42,6 @@ const BottomSheet = ({ forwardedRef, title, description, children, onDismiss, te
       enablePanDownToClose
       backdropComponent={renderBackdrop}
       handleIndicatorStyle={styles.handle}
-      onChange={handleChangePosition}
     >
       <BottomSheetScrollView handleContentLayout={handleContentLayout} testId={testId}>
         {title && <Text style={styles.title}>{title}</Text>}
